+++ stdnet/utils/__init__.py
@@ -1,6 +1,3 @@
-import time
-from datetime import datetime
-
 from stdnet.lib.py2py3.py2py3 import *
 if ispy3k:
     import pickle
@@ -15,6 +12,7 @@ else:
 from .jsontools import *
 from .populate import populate
 from .fields import *
+from .dates import *
     
     
 class NoValue(object):
@@ -26,27 +24,4 @@ class NoValue(object):
 
 novalue = NoValue()
 
-
-
-def date2timestamp(dte):
-    return time.mktime(dte.timetuple())
-
-
-def timestamp2date(tstamp):
-    "Converts a unix timestamp to a Python datetime object"
-    return datetime.fromtimestamp(float(tstamp))
-
-    
-def todatetime(dt):
-    if isinstance(dt,datetime):
-        return dt
-    else:
-        return datetime(dt.year,dt.month,dt.day)
-    
-    
-def todate(dt):
-    if isinstance(dt,datetime):
-        return dt.date()
-    else:
-        return dt